add copy constructor of variable
