+++ src/variables/variables_internal.h
@@ -20,6 +20,7 @@ t_variable	*variable_create(char *name, char *value);
 void        *variable_copy(void *data);
 
 // utils.internal.c
+char        *make_envstr(char *name, char *value);
 void		variable_destroy(void *data);
 int			add_new_variable(t_context *ctx, char *name, char *value);
 int			update_variable(t_variable *var, char *value);
